requirements.txt
boto3~=1.20.26
boxsdk~=2.13.0
dnspython~=2.1.0
gensim~=3.8.3
pdf2image~=1.16.0
Pillow~=8.4.0
pymongo~=3.12.0
pytesseract~=0.3.7
python-dotenv~=0.19.0
yake~=0.4.8
